✨ Add vibration to the end of Hit Timer
